#4953. Suburb autocomplete with jquery.ui. Site page PageType section.
git-svn-id: https://svn.ish.com.au/ish/willow/trunk@11299 fb2e276d-fae9-0310-b578-b71ac85840b9

+++ frameworks/common-model/src/main/java/ish/oncourse/model/auto/_PostcodeDb.java
@@ -42,11 +42,11 @@ public abstract class _PostcodeDb extends CayenneDataObject {
         return (Double)readProperty("lon");
     }
 
-    public void setPostcode(Integer postcode) {
+    public void setPostcode(Long postcode) {
         writeProperty("postcode", postcode);
     }
-    public Integer getPostcode() {
-        return (Integer)readProperty("postcode");
+    public Long getPostcode() {
+        return (Long)readProperty("postcode");
     }
 
     public void setState(String state) {
